--- website/src/utils/Error/error.css
@@ -1,78 +1,0 @@
-/*
-Copyright 2023 The Vitess Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-.error {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin-top: 3rem;
-}
-
-.error div {
-  margin: 1rem;
-  display: flex;
-  align-items: center;
-  flex-direction: column;
-}
-
-.goHome {
-  padding: 10px 22px 10px 22px;
-  margin: 20px;
-  font-size: 1rem;
-  border-radius: 1rem;
-  background-color: #e77002;
-  color: white;
-  border: none;
-  cursor: pointer;
-  transition: 300ms ease-in;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  column-gap: 10px;
-}
-
-.errorImg{
-  padding-left: 30px;
-}
-.error div img {
-  padding: 10px;
-  width: 260px;
-  height: 225px;
-  animation: translateUpDown 2s infinite alternate;
-}
-
-.error h1 {
-  font-size: 3rem;
-  font-weight: 700;
-}
-
-.error h2 {
-  font-weight: 500;
-  text-align: center;
-}
-
-@keyframes translateUpDown {
-
-  0%,
-  100% {
-    transform: translateY(0);
-  }
-
-  50% {
-    transform: translateY(-0.7cm);
-  }
-}
